requirements.txt
selenium
pandas
pytest
undetectable-chromedriver
python-dotenv  # Add this line
